+++ Source/Libraries/CorruptCore/CorruptCore.csproj
@@ -100,6 +100,7 @@
     <Compile Include="EventWarlock\WarlockAction.cs" />
     <Compile Include="EventWarlock\WarlockActions\WarlockActionEcho.cs" />
     <Compile Include="EventWarlock\WarlockConditions\FirstEqualsSecond.cs" />
+    <Compile Include="Memory\FileInterface.cs" />
     <Compile Include="Memory\MemoryDomainProxy.cs" />
     <Compile Include="Memory\MemoryDomains.cs" />
     <Compile Include="Memory\MemoryInterface.cs" />
